renamed src/utils to src/main_utils with consequential changes
