+++ src/main_utils/enums.py
@@ -5,4 +5,5 @@ class ThreadState(Enum):
     AWAITING_LOGS = auto()
     LOGS_RECEIVED = auto()
     NO_OLYMPUS_LOGS = auto()
+    DCS_LOG_RECEIVED = auto()
     CLOSED = auto()
